requirements.txt
FastAPI==0.63.0
FastAPI-SQLAlchemy==0.2.1
psycopg2-binary==2.8.6
pydantic==1.7.3
pytest==5.4.2
python-dotenv==0.15.0
requests==2.7.0
SQLAlchemy==1.3.20
uvicorn==0.13.3
